requirements.txt
python-telegram-bot==12.2.0
Pyrogram==0.16.0
jdatetime==3.6.2